add service / route for claim details
