+++ Fmas12d.Ui/src/app/interfaces/state.ts
@@ -1,3 +1,5 @@
+import { SortDirection } from '../directives/table-header-sortable/table-header-sortable.directive';
+
 export interface State {
   page: number;
   pageSize: number;
